Add 'Order Completed' support for Mixpanel Actions
